DBAL changes for gNMI subscription
1) Introduced the OnChange cache in the DB object. Cache is maintained
as map[string]db.Value, where map key is the redis key.

OnChange cache is not enabled by default. First the DB instance should
be created by setting IsEnableOnChange=true in the db.Options object.
Then the cahce needs to be enabled for specific tables by calling
DB.RegisterTableForOnChangeCaching() for each of them.
DB.OnChangeCacheUpdate() and DB.OnChangeCacheDelete() functions can be
used to refresh the cache contents.

2) Enhanced DB.GetEntry() to return value from the OnChnage cahce if
the key exists in the cache. Otherwise the value will be read from
redis and cache will be updated.

3) Add Key.Matches() function to match a key to a redis glob pattern

4) Add ScanCursor object which enables redis SCAN operation